Added propper caching!

+++ CMH.Priority/Controller/TestController.cs
@@ -6,7 +6,7 @@ using CMH.Common.Extenstion;
 using CMH.Common.Message;
 using CMH.Data.Repository;
 using CMH.Common.Variable;
-using CMH.Common.Util;
+using System.Threading;
 
 namespace CMH.Priority.Controller
 {
@@ -15,20 +15,17 @@ namespace CMH.Priority.Controller
     public class TestController : ControllerBase
     {
         private readonly HttpClient _functionHttpClient;
-        private readonly ICacheManager _cacheManager;
         private readonly IDataSourceRepository _dataSourceRepository;
         private readonly ServiceBusClient _serviceBusClient;
         private readonly ServiceBusAdministrationClient _serviceBusAdministrationClient;
 
         public TestController(
             IHttpClientFactory httpClientFactory,
-            ICacheManager cacheManager,
             IDataSourceRepository dataSourceRepository,
             ServiceBusClient serviceBusClient,
             ServiceBusAdministrationClient serviceBusAdministrationClient)
         {
             _functionHttpClient = httpClientFactory.CreateClient("Function");
-            _cacheManager = cacheManager;
             _dataSourceRepository = dataSourceRepository;
             _serviceBusClient = serviceBusClient;
             _serviceBusAdministrationClient = serviceBusAdministrationClient;
@@ -42,8 +39,12 @@ namespace CMH.Priority.Controller
             var queues = _serviceBusAdministrationClient.GetQueueNamesAsync("").Result;
             foreach (var queue in queues)
             {
-                await _serviceBusAdministrationClient.DeleteQueueAsync(queue);
-                await _serviceBusAdministrationClient.CreateQueueAsync(queue);
+                var queueProperties = await _serviceBusAdministrationClient.GetQueueRuntimePropertiesAsync(queue);
+                if(queueProperties.Value.TotalMessageCount > 0)
+                {
+                    await _serviceBusAdministrationClient.DeleteQueueAsync(queue);
+                    await _serviceBusAdministrationClient.CreateQueueAsync(queue);
+                }
             }
 
             //Reset process
@@ -52,9 +53,6 @@ namespace CMH.Priority.Controller
                 _functionHttpClient.BaseAddress = new Uri(string.Format(_functionHttpClient.BaseAddress?.ToString() ?? "", "process/reset"));
                 await _functionHttpClient.PostAsync("", null);
             } catch { }
-
-            //Reset cache
-            _cacheManager.Clear();
 
             //Post messages
             var maxMessageBatch = 500;
